+++ src/com/cloudogu/ces/dogubuildlib/EcoSystem.groovy
@@ -109,8 +109,12 @@ class EcoSystem {
         vagrant.ssh "sudo cesapp build ${doguPath}"
     }
 
-    void purgeDogu(String doguName) {
-        vagrant.ssh "sudo cesapp purge ${doguName}"
+    void purgeDogu(String doguName, keepConfig = false) {
+        if (keepConfig) {
+            vagrant.ssh "sudo cesapp purge --keep-config ${doguName}"
+        } else {
+            vagrant.ssh "sudo cesapp purge ${doguName}"
+        }
     }
 
     void installDogu(String doguFullName) {