+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Parameter for EcoSystem.purgeDogu function to keep dogu config; #26
 
 ## [v1.5.1] - 2021-11-17
 ### Fixed